17266: Fixes integration test.
One cypress test started failing because instead of using visit() now the
tests navigate to collection through the app router, so previous state is
kept. The problem was that on collection updates an action was dispatched
to update the parent project of the collection. This caused issues with the
details panel incorrectly loading the project's details instead of keeping
the collection's data on it.

Also, updated yarn.lock after the cypress 6.x update.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas@di-pentima.com.ar>