16848: Fixes unit tests.
DetailsAttribute component needed a mocked store to be able to be tested
inside TokenDialog. Extracted the layout related code to a simpler component
so that can be used directly and tested is easier.
Also, fixed a couple of assertions on the token dialog tests.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas@di-pentima.com.ar>