Tweak version-at-commit.sh for interim releases (X.Y.Z.A)
Assign correct intended dev version when there is a X.Y.Z.A tag (not
just X.Y.Z) in the history.

As of this writing we don't have tags like that in the arvados repo,
but we do have tags like that in the workbench2 repo, so this commit
is just about synchronizing the script for consistency between the two
repos.

refs #17512

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>