17389: Adds X-Keep-Storage-Classes-Confirmed header to responses on success.
Also, don't treat partial success (ie: replicas written > 0) as success to
let the client decide what to do, as this would require a GoSDK API change on
Put methods, or the use of a special new method just for keepproxy that
returns fulfilled storage classes information.
In the case of partial successes from the client point of view, the only
thing that a client can do is retry the request with the same keepproxy, and
that would render the same result.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas.dipentima@curii.com>