15430: Fixes updated wb1 test that was migrated from using '@@' to 'ilike'.
It seems that the ordering isn't the same when using both operators as one
pipeline instance object wasn't getting listed after migration. Had to both
change the created_at and uuid fields to make it appear on the all_processes
page when passing 'ilike'.

This shouldn't be a problem, as we're actually using 'ilike' on the real wb1
page, it was just the test what was outdated.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas.dipentima@curii.com>