18128: Adds Multi Panel View (MPV for short) component family.
MPVContainer aims to be a drop-in replacement for <Grid container...>
that will handle its children's visibility and pass to them additional
props so they can optionally offer UI elements to the user.

MPVPanelContent aims to be a drop-in replacement for <Grid item...>
in case there's a need for special layout inside a container. This
component will forward the additional props to its children.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas.dipentima@curii.com>