17492: Clean up bundler version check/install.
Apparently gem installs bundler in $HOME, not in $GEM_HOME like we
expected.  The broken code continued working as long as it was using a
pre-existing run-tests temp dir that also installed a version of
bundler in $GEM_HOME because bundler was listed in a Gemfile.lock.

Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tom@curii.com>