18560: Improves the property form's autocomplete feature.
Instead of presenting the whole list of terms and their synonymns to the user,
only show the preferred term (the first defined label in the vocabulary) and
if the user starts typing, also add the synonyms to every preferred term.
This way the list gets a lot smalle and easier to browse.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas.dipentima@curii.com>