Remove the pre-seeding of gems needed by arvados-api-server and arvados-workbench from the package build Dockerfiles.
This step was doing a git clone from main. This can cause issues when
the packages being built are for another branch. For example, when the
Gemfile in main has gems with a dependency on a version of Ruby that is
newer than the one installed in these Dockerfiles on the branch being
built, which is what happened today.

The git clone should really be for the commit being built, but that
would make the image specific to that build, which means we might as
well not pre-seed.

The pre-seeding of these gems was of dubious utility anyway, since we
rebuild the package build images on every package build step invocation
on the CI server.

refs #18766

Arvados-DCO-1.1-Signed-off-by: Ward Vandewege <ward@curii.com>