16115: Sets expiration date's minutes & seconds to zero.
Default expiration date is now set in a way that it gives the sharing link
at least 1:01h of lifetime and up to 2h. The user can select other dates
with a minimum step resolution of 1h.
Also, the "Create sharing URL" button won't be enabled if the selected
date is in the past. This is done this way because the Time picker doesn't
seem to support a "minDate" (or minHour?) property.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas.dipentima@curii.com>