18975: Fixes log viewer's follow mode when maximized.
Added a 10% of scroll margin so that it's guaranteed that follow mode is
enabled no matter the log viewer's height.
The side effect is that the scrolling gets a "snap effect" when the user
gets close to the end of the log.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas.dipentima@curii.com>