18842: When starting up the disk cache, map in everything
When a process starts, the first thing it will do is map in the
existing blocks and prune any excess from the cache.

If there are multiple processes using the same cache dir, at worst
you'd end up with N*M usage where N is the number of processes and M
is the cache limit.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>