18797: Update style of arvados.retry docstrings.
I wrote these docstrings as a demonstration of new style rules for our
Python SDK. It is real, useful documentation, as well as a good
illustration of how to follow the new style guide, so we might as well
merge it.

This is primarily a style update to the existing docstrings. In places
where I've actually changed the meaning, I believe the new meaning is
more correct. In particular, RetryLoop's "attempts" count the number of
saved results, which may or may not be the same as the number of times
the loop has run depending on what the body does.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>