19886: Slim down API documentation about container logs
While the API server will let you set this field to a UUID, that
behavior is already poorly supported, and on reflection we're not sure
we want to support it. Document this the way crunch-run uses it and
avoid presenting too much detail for most readers.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>