20487: loadProject doesn't call openProjectPanel
The bug was was that loading the registered workflow panel would call
loadProject() which would call openProjectPanel() which would dispatch
REQUEST_ITEMS for the project (which is 100% wasteful, because we're
not displaying the project contents).

However, the results from REQUEST_ITEMS (which excludes some fields in
the standard query) would replace the result that previously got the
complete workflow object.  As a result the "definition" field went
away, which cause the inputs/outputs panels to go blank.

This switches the navigation action to call "openProjectPanel", and
have that begin by calling "loadProject".  As a result, other places
that call loadProject no longer call openProjectPanel.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>