19793: Make browser back button work on nav to PDH
When navigating to a collection by PDH, if there is only one match, it
the portable data hash page automatically directs the user to the
appropriate collection page.  However, if the user pushes the back
button, they go back to portable data hash page which automatically
takes them to the collection page again.  This effectively breaks the
back button.

This fixes the bug by replacing (instead of "pushing") the top of the
location history with the collection page, so that the back button
from the collection page takes the user to their previous page without
revisiting the portable data hash page (so they are not endlessly
redirected).

This commit message is longer than the actual code fix.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>