20839: Mark deprecated PySDK functions as such
This is in preparation to remove these from the SDK in Arvados 3.0.

These changes exacerbated the circular import between the top-level
arvados and util, which had knock-off effects on other import
statements, causing them to see the world differently (like whether
`arvados.api` is a module or function). Most of the import-related
shuffling is meant to mitigate all that.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>