21028: Build Ruby from source on rocky8
Without this, RVM downloads a binary Ruby with a bunch of configuration
options turned on. This can lead to problems building extensions when
Ruby thinks a certain feature is supported but our Docker image doesn't
have the headers to support it. Right now we're seeing that in the zlib
gem: it has extra code to turn on when Ruby supports Valgrind, but we
don't have the Valgrind headers necessary to build that code.

Closes #21028.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>