21077: Add explicit START/STOP_WORKING to loading functions
Previously, progressFn meant that any time there was an active API
request, it would spin the progress bar.  To allow for background
requests (where we don't want it to spin because it is
distracting/confusing to the user), we need to remove that default
behavior.  As a result, to provide feedback that something is
happening, functions need to explicitly set START_WORKING and
STOP_WORKING.  This was implemented inconsistently, because the
default processFn behavior tended to cover a lot of cases.

This commit adds START/STOP to all the major panel loading functions
in the UI.  This provides better user feedback overall, because the
spinner now more consistently covers the entire loading
process (across multiple API calls and async behavior), instead of
just individual API calls.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>