21136: Use typing generics for builtins in api_resources
The immediate problem this solves is that using `list` unqualified as an
annotation causes pdoc to resolve it as the current class' `list`
method, which is not correct.

We could also specify `builtins.list`, but since it doesn't support
subscripting in Python 3.8, that generates a lot of noisy warnings right
now. So this solution is better for now, and we'll probably migrate to
`builtins.list` in the future.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>