21132: Annotate API methods as returning a request object
This is a factual correction. Before this, it looked like the API
methods communicated with the API server and returned a response
directly, which is not true.

Unfortunately pdoc doesn't render the documentation for the
ArvadosAPIRequest.execute override. That's frustrating. I looked for
ways to coerce this but didn't find any. Hopefully the class docstring
gives users enough information about what they need to do. pydoc renders
it as expected, so I still think it's worth keeping.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>