21137: Remove id_token_hint from controller RP-initiated logout
In the current Arvados stack, it is not possible to provide the ID token
this parameter expects, because we do not retain it anywhere. The user's
authorization token will either be an OIDC access token or an Arvados
token generated when they logged in, so neither is suitable for use as
id_token_hint.

The spec says this parameter is RECOMMENDED but not REQUIRED, so
omitting it is the simplest option that should still meet functional
requirements.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>