21146: Add last_log_id argument to PollClient.subscribe
This helps avoid a crash if `subscribe` returns a `PollClient` and then
the user calls `subscribe` with two arguments. I don't know whether
ignoring `last_log_id` is *correct*, but that's the current behavior,
and anything else is out of scope for the current ticket.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>