Use full paths for API documentation links
Most of the cross-reference links in our documentation point to a
specific page. These links are unusual in that they point to a
directory. Lately the linkchecker has been intermittently complaining
about them:

  URL        `file:///tmp/workspace/developer-run-tests-doc-and-sdk-R/doc/.site/api'
  Name       `API documentation'
  Parent URL file:///tmp/workspace/developer-run-tests-doc-and-sdk-R/doc/.site/admin/scoped-tokens.html, line 306, col 576
  Real URL   file:///tmp/workspace/developer-run-tests-doc-and-sdk-R/doc/.site/api/
  Warning    [file-missing-slash] Added trailing slash to
             directory.
  Result     Valid: directory

This causes a test failure. Spell out the rest of the link to prevent
that. No issue #.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>