20846/21230: Remove Python transitional metapackages from the package build
The build process for this is relatively expensive (we set up a
dedicated virtualenv for each one) and not PEP 668-compliant (it tries
to upgrade packages globally). We don't need these packages anymore for
the Arvados 3.0 release. Stop building all of them.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>