20318: Fix settled condition.
Intention was to stop iterating when *all* of the keepstore servers
need 0 pulls, but the code as written stops on the N>1st iteration
when *any* of the keepstore servers need 0 pulls. This caused a false
negative when it took more than one iteration for keepstore to
complete the requested pulls.

Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tom@curii.com>