21700: Better isolate Bundler from the build source
This helps ensure that when we build the RailsAPI package with Bundler,
we don't accidentally pick up settings or files from the developer's
machine. This shouldn't matter when we build packages on Jenkins but
this makes it easier for developers to test package builds on their own
systems.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>