21700: Remove RVM support
We haven't reliably tested this in a long time, and it's a lot of code
for no benefit.  In general we want to support the versions of Ruby that
come with the distributions we support. Anybody in a very weird case can
install a supported version of Ruby from source.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>