21700: Standardize on Bundler~>2.4.0
These are the last versions to support Ruby 2.7, which we're still
supporting as I write this.

Our usual install recipe is:

  gem install --conservative --version '~> 2.4.0' bundler

In environments where we *know* we have Ruby 3, that becomes:

  gem install --conservative --version '>= 2.4.0' bundler

As long as we have something at least as recent as 2.4.0, that should be
enough for function. This recipe meets our functional needs while
striking a balance between being easy to implement and avoiding
redundant installs.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>