21207: Remove Python slow/short tests support
After the previous commit there are no more users of this, and this
interface will no longer be supported after our migration to pytest. If
we want this functionality in the future we should set up a custom
"slow" mark in pytest and use pytest's built-in mark filtering.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>