21020: Note configuration differences when $XDG_CONFIG_HOME is set
Customizing $XDG_CONFIG_HOME is expected to be uncommon. Users who do it
are expected to know what they're doing. Therefore we do not provide
complete instructions for how to adapt, we just note it and expect them
to be able to follow high-level instruction.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>