21388: Remove packaging commitment
First of all, it's out-of-date. Our current *aim* is to support the
current and previous stable release of Debian; the current and previous
LTS release of Ubuntu; and current RHEL. In addition to that, the
vagaries of release cadence and supporting newer distros means we're not
always up-to-date with that target.

Just remove this sentence. It doesn't add anything compared to the table
above, which is authoritative.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>