21388: Delist CentOS 8 support
CentOS 8 went EOL at the end of 2021: <https://www.centos.org/centos-linux-eol/>

Given CentOS' change in development direction, we are not aware of any
users interested in CentOS support, and don't anticipate any for the
foreseeable future. De-list support for it. The packages might happen to
work, but we don't support them.

Our RPM URL still contains "CentOS" in the path. Cleaning that up is a
separate infrastructure task.

List "Red Hat" as the first supported distro since it's the biggest one,
and the one people might be skimming for.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>