21572: Remove GPG fingerprint
This information is not useful to users if we don't tell them how to
verify the fingerprint themselves.

And, that's nontrivial to do on Red Hat distros because yum downloads
and stores the key automatically.

And, there's not much reason for users to trust the fingerprint that
comes from https://doc.arvados.org more than they trust the key itself
from https://public.arvados.org.

So just remove this.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>