21906: Improve Rails postinst unconfigured reporting
The immediate bug this is fixing is: on Red Hat, if you installed the
package before a web server, it would not install /etc/arvados (because
it didn't know who should own it), but then tell you to edit
/etc/arvados/config.yml.

This commit improves that reporting, so it gives you a more specific
reason that configuration stopped, and next steps for the administrator
to take. Admittedly, I did quite a bit of work to get there, but here we
are.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>