22003: Workbench2 and keep-web: better interoperability with redirect
keep-web: When sending an unauthenticated browser client to a redirect
to Wb2, encode the target URL path in the query part of the redirection
URL in the `Location` header. This avoids possible corrupted header and
confused client.

Workbench2:

- In redirection handler, more robust handling of the input target-path
  passed by the URL query part.
- In the "copy link to clipboard" action in the files panel of a
  collection view, when creating the URL for clipboard, better emulate
  the server-generated redirect URL (see above).

Overall, when working with redirects (either generating redirect URLs or
handling them on the client side), we're better prepared for paths that
may contain special characters.

Arvados-DCO-1.1-Signed-off-by: Zoë Ma <zoe.ma@curii.com>