21935: Move arvados.http_to_keep under arvados._internal
This functionality could be useful to provide to people writing their
own tools. However, because of the way it was extracted out of
arvados-cwl-runner, the API wasn't really designed to be public, and has
already had one breaking change since it was introduced.

Make the current module internal so the API doesn't ossify. We can
always develop a public API for it in the future.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>