21909: Add a basic build system to R SDK
The R SDK includes committed code and documentation that are generated
from other sources, mainly the discovery document and inline
documentation. Remove those and introduce a Makefile with recipes to
build them all. This will make it easier to keep the R SDK in sync with
Arvados API changes in the future.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>