22124: Reject requests with double slash, except GET //locator.
Previously we responded to "PUT //locator" with a 301 to a cleaned
path, which did not work because 301 tells clients to repeat the PUT
request as a GET.  Better to just respond 400, since this means the
client is broken or misconfigured.

There may be some clients that have been working successfully with the
301 redirect behavior because they only do GET and HEAD requests.  To
avoid breaking these, we continue to accept "GET //locator" -- but now
we just return the data instead of a redirect to "GET /locator".

Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tom@curii.com>